replace single value
